// src/generated/models/BC_Announcements_SchemaModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface field_2Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface field_3Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface field_8Value {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface TargetSchoolsValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface BC_Announcements_Schema {
  ID?: number;
  Title?: string;
  field_1?: string;
  "field_2#Id"?: unknown;
  "field_2@odata.type"?: string;
  field_2?: field_2Value;
  "field_3#Id"?: unknown;
  "field_3@odata.type"?: string;
  field_3?: field_3Value;
  field_5?: string;
  field_6?: string;
  field_7?: boolean;
  "field_8#Claims"?: unknown;
  "field_8@odata.type"?: string;
  field_8?: field_8Value;
  field_9?: string;
  "TargetSchools#Id"?: unknown;
  "TargetSchools@odata.type"?: string;
  TargetSchools?: TargetSchoolsValue;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  Author?: AuthorValue;
  "Editor#Claims"?: string;
  Editor?: EditorValue;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: Record<string, unknown>;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: Record<string, unknown>;
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: unknown;
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}